fix(frontend): 投稿を削除したときに更新されるように変更
